+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+### Added
+- A `--build` option to the init command, to be able to automatically bootstrap a database in genome build 38
+
 ## [4.7.2] - 2024-12-10
 - Refactored to allow later python versions than 3.8 (pkg_resources to importlib, os.path and path.py to pathlib, distutils to shutil) [#262](https://github.com/Clinical-Genomics/chanjo/pull/262)
 - Updated actions to use python 3.11, add changelog reminder, add linting, format for linters [#266](https://github.com/Clinical-Genomics/chanjo/pull/266)